[feat] #1 디자인 모듈 연결 및 적용 완료
