+++ app/src/main/java/com/desserttime/desserttimeapp/MainActivity.kt
@@ -11,7 +11,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
-import com.desserttime.desserttimeapp.ui.theme.DessertTimeAppTheme
+import com.desserttime.design.theme.DessertTimeAppTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
