+++ app/src/main/res/drawable-anydpi/ic_launcher.xml
