+++ app/src/main/res/values/themes.xml
@@ -6,8 +6,8 @@
 <!--    splash-->
     <style name="Theme.DessertTimeApp.Splash" parent="Theme.SplashScreen">
         <item name="windowSplashScreenAnimationDuration">1000</item>
-        <item name="windowSplashScreenBackground">@color/purple_200</item>
-        <item name="windowSplashScreenAnimatedIcon">@drawable/ic_launcher_background</item>
+        <item name="windowSplashScreenBackground">@color/bg_screen</item>
+        <item name="windowSplashScreenAnimatedIcon">@drawable/ic_screen_logo</item>
         <item name="postSplashScreenTheme">@style/Theme.DessertTimeApp</item>
     </style>
 </resources>
