--- design/src/main/java/com/desserttime/design/extensions/Density.kt
@@ -1,10 +1,0 @@
-package com.cmc.curtaincall.common.designsystem.extensions
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.platform.LocalDensity
-import androidx.compose.ui.unit.Dp
-
-@Composable
-fun Dp.toSp() = with(LocalDensity.current) {
-    this@toSp.toSp()
-}
