+++ design/src/main/java/com/desserttime/design/theme/Type.kt
@@ -1,38 +1,33 @@
 package com.desserttime.design.theme
 
-import androidx.compose.runtime.Immutable
-import androidx.compose.runtime.staticCompositionLocalOf
+import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.font.FontFamily
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.sp
 
-@Immutable
-data class DessertTimeTypography(
-    val h1: TextStyle,
-    val h2: TextStyle,
-    val subTitle1: TextStyle,
-    val subTitle2: TextStyle,
-    val subTitle3: TextStyle,
-    val subTitle4: TextStyle,
-    val body1: TextStyle,
-    val body2: TextStyle,
-    val body3: TextStyle,
-    val body4: TextStyle,
-    val body5: TextStyle,
-    val caption: TextStyle
-)
-
-val LocalCurtainCallTypography = staticCompositionLocalOf {
-    DessertTimeTypography(
-        h1 = TextStyle.Default,
-        h2 = TextStyle.Default,
-        subTitle1 = TextStyle.Default,
-        subTitle2 = TextStyle.Default,
-        subTitle3 = TextStyle.Default,
-        subTitle4 = TextStyle.Default,
-        body1 = TextStyle.Default,
-        body2 = TextStyle.Default,
-        body3 = TextStyle.Default,
-        body4 = TextStyle.Default,
-        body5 = TextStyle.Default,
-        caption = TextStyle.Default
+val Typography = Typography(
+    bodyLarge = TextStyle(
+        fontFamily = FontFamily.Default,
+        fontWeight = FontWeight.Normal,
+        fontSize = 16.sp,
+        lineHeight = 24.sp,
+        letterSpacing = 0.5.sp
     )
-}
+    /* Other default text styles to override
+    titleLarge = TextStyle(
+        fontFamily = FontFamily.Default,
+        fontWeight = FontWeight.Normal,
+        fontSize = 22.sp,
+        lineHeight = 28.sp,
+        letterSpacing = 0.sp
+    ),
+    labelSmall = TextStyle(
+        fontFamily = FontFamily.Default,
+        fontWeight = FontWeight.Medium,
+        fontSize = 11.sp,
+        lineHeight = 16.sp,
+        letterSpacing = 0.5.sp
+    )
+    */
+)
